[davinci] Added scaling support to video layer using the Davinci Resizer.
Surface buffers must be allocated in devmem as the fb memory of the layer
is used to store the resized data. Applications can set source and destination
rectangles as with other scaled layers, but a Flip() is always required to
show updates even in a single buffer mode.

Minor fixes and cleanups.