[surfaces] Added Surface Pool Bridge[tm] technology to speed up transfer between pools.
New API can be implemented to do optimized transfer from one surface pool to another.

CheckTransfer() is called to probe each registered CoreSurfacePoolBridge.

CoreSurfacePoolTransfer is allocated with one or more rectangles to be transfered.

StartTransfer() and FinishTransfer() are used to actually execute the transfer
which could be queued, but is synchronous at the moment (Start+Finish).