modules: Don't touch the module entry after dlclose()
When a module was unloaded with dlclose() the destructor function
(direct_modules_unregister() for the input drivers) was called. The
dtor freed the module entry but the code calling dlclose() proceeded to
fiddle with the already freed entry afterwards. Reorganize the code
slightly to avoid the problem and add a few comments to remind people
what may happen when dlclose() is called.

The problem sometimes manifested itself as a segfault when loading the
modules. Suppressing all but the required input modules was an
effective workaround for the problem.