font: introduce new font attribute DFFA_NOBITMAP
Some truetype fonts include bitmap strikes for certain sizes. A new font
attribute is introduced to the DFBFontAttributes enum. It maps to
the FT_LOAD_NO_BITMAP attribute in the freetype2 renderer. This
tells the renderer to always provide an anti-aliased rendering,
ignoring bitmap strikes embedded in the font. For bitmap-only fonts this
flag is ignored.