vt/keyboard: Flush console input when not used
To keep the kernel happy the console buffer must be kept from filling
up. If the buffer is sufficiently full the kernel stop sleeping and
instead wakes up every tick to do something. This problem occurs when
the keyboard driver is not used and the linuxinput driver does not grab
the devices. Add a new thread whose job is to keep flushing the buffer
when the keyboard driver is not there to consume the data.