[colorize windows] allow the window config value "color" to colorize the window
This parameter was reserved when a window was created with DSCAPS_COLOR:
it would show the window with this color only (single color, no surface needed).
Now it is also active without DSCAPS_COLOR:
if you specify a color with a non-nul alpha value, the window
will be blended to the screen using DSBLIT_COLORISE: this multiplies
the window content with this color.