debug: prevent deadlock in stack trace with SaWMan active
This can only happen when a stack trace is printed.
Previously, the debug log system was locked during this operation.
When a symbol needed to be looked-up, "nm" would be started which
could lead to a "process_added" debug print in SaWMan;
this would lead to deadlock since the log system was locked already.

Solution is to accumulate all statements (incl. symbol lookup)
and lock-print-unlock in one go.