input: initialize linux_input key states
IDirectFBInputDevice caches the key states (up/down) by keeping track
of press and release events. This doesn't take into account that a key
may be pressed at the time the device is opened or that a key might
have changed state while the device was in suspend.

This change introduces code to the linux_input driver that queries
the keystates when the event thread is started and synthesizes key
events for all keys depending on the key state as reported by the
linux input layer.

Signed-off-by: Sven Neumann <s.neumann@raumfeld.com>