font: drawstring clipping
A check is done before glyph blitting to ensure that at
least some portion is within the clipping rectangle.
The check is wrong as it assumes the glyphs are referenced
by the top left of the bounding box instead of the baseline origin.

Thanks to Tim Ross for the patch.