Fix for showing unitialized buffer and superfluous clear at startup.
Fixes display layer window stack repainting to NOT be performed
when it is frozen.  This prevents a display layer surface flip
from being done and showing an uninitialized buffer when the
IDirectFB::GetDisplayLayer function is called with an associated
no-init-layer directfbrc command (without init-init[=0-15]).

Fixes IDirectFBDisplayLayer::GetSurface function to only perform
single buffered display layer clearing when the region is frozen
and to render the background if it is configured
(each time GetSurface is called the surface should not be cleared).
Also added support for this behavior when the cooperative level is
DLSCL_ADMINISTRATIVE.

Thanks to Timothy Strelchun for the patches.