core window stack: Fixed _dfb_windowstack_inputdevice_listener race condition with layer context
Fixed a race condition when _dfb_windowstack_inputdevice_listener
begins using a layer context, that is then destroyed before the
listener has finished using it.  This problem occurs because the
layer context's reference count was never incremented to account
for the listener's reference to it that it stored and subsequently
used.  Rather than establishing a reference count cycle (with the
window stack having a reference to the layer context which has a
reference to the window stack for which additional complexity
would be needed to break), the listener now just increments the
layer context's reference count prior to usage and then decrements
it.

Patch by: From: Timothy Strelchun <Timothy.Strelchun@Intel.Com>