Revert "If a dlopen()'ed module did not register itself, try to dlsym() and call"
This reverts commit f1a8923dc6a4499e0bb24814e3e251753850043b.

Currently, DirectFB's lib/direct/modules.c will try to call a function with
a name of shared_library_name with lib and .so removed in case the module
did not have an __attribute__((constructor)) function. This works for many
cases of the existing modules in the DirectFB source tree, but not for all:

1) the constructor function name of many (but not all) modules has had the
string "_ctor" added a while ago.

2) some modules don't fit into this naming scheme (all which use the
DIRECT_INTERFACE_IMPLEMENTATION macro to implement an interface), as they
use different cApITAliSaTION rules for the function type name vs. the
shared library name (e.g. IDirectFBImageProvider_GIF_ctor() vs.
libidirectfbimageprovider_gif.so).

These two actually mean that trying to call a 'default' constructor function
after dlopen() has not been useful for quite a while.