gfxdrivers/gl: Added experimental fragment shader for source colorkeying.
This boosts framerate in df_andi from 12fps to 245fps on my nvidia ion box.

Compiling the shader is done on-demand. It should be done in init_driver()
but OpenGL is not ready to compile shaders there.

If compilation of the fragment shader is successfull, add DSBLIT_SRC_COLORKEY
to the supported blittingflags.

The downside of on-demand compilation is that source colorkeying will never
be reported in the CardCapabilities. So applications querieing them will get
always be reported source colorkeying as being unsupported.