sh772x: Register phys memory with UIOMux
This is patch significantly improves the performance on sh772x devices.

The sh772x gfx code uses an external library (libshbeu) to access the
BEU hardware. In order for this library to understand that the virtual
addresses passed to it are for contiguous memory, the memory must be
registered with another external library (UIOMux).

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>
(cherry picked from commit b37828dd5dcd970a207b81fa771c71acb3b5aa26)