vdpau: New X11/VDPAU based system module derived from original X11 module.
The module also creates an X11 window, but instead of using XPutShmImage
or similar it creates a VdpPresentationQueueTarget and VdpPresentationQueue.

There's also a surface pool creating VDPAU Output Surfaces for use by layers
via VdpPresentationQueueDisplay and later for accelerated operations using
the surface rendering functions of VDPAU.

This is the initial working version, successfully tested with df_dok and df_andi.

The code still needs to be cleaned up a lot, multiple layers need to be
composited using VdpMixer functions and special video layer support for playback
will be added, so that VdpOutputSurfaces from Xine or GStreamer can be pushed
into DirectFB without copying their pixels, compositing them with the other
surfaces from the graphics layers, keeping the video timestamps as master clock.