MESA: added simple VT code
The code is basically a copy of the fbdev vt code minus the fb specific stuff.
This should be merged when mesa leaves the experimental status.

Starting a DirectFB application from X  will still make it crash after we
switch back to X's vt. I guess we need to add some cleanup code before doing
he switch back to the original vt.

Still this is of value for debugging, because when a DirectFB application
gets started from a non-x vt, debugging output will be visible instantly after
the application quits. It was annoying to switch so some vt and then
searching for the original vt.