Proxy: Implement asynchronous flipping control, e.g. to avoid input lags
For each Flip an asynchronous FlipNotify request is sent from dispatcher
to requestor instance. For older servers a workaround using a hidden dummy
window and an event buffer is implemented, but at the moment it creates
a window/buffer per surface.

With the asynchronous flipping control, the client waits upon Flip when
there are already two Flips on their way...