IDirectFB: Get rid off idirectfb_singleton, instance per client, client mode/pixelformat/window config.
This also fixes the primary window ref recently broken. The original fix
was wrong, the real issue was that IDirectFB did not get destructed as
it was shared with other clients and could not unref the primary window.

The run time options 'mode', 'pixelformat' and 'force-windowed' are handled
on client side as well now.