Core: Add FusionCalls for CoreLayerContext and CoreWindowStack.
Initial function is CoreLayerContext_CreateWindow used to create a
window from slave. The call returns the id of the new window object
which is looked up and refed in the slave.

At the moment this is leaking a ref as the master created the object,
but we're going to remove the initial ref up in the object creation
and let the slave do the first ref up, otherwise call object destructor
after a timeout in case the slave does not pick up the new object.