SecureFusion: Final major step to get SaWMan working with secure fusion
Add SaWManManager flux to implement the public ISaWManManager interface for the application manager
which could be a slave process without write access to the SaWMan structures. This includes
QueueUpdate(), ProcessUpdates(), CloseWindow(), InsertWindow(), RemoveWindow(), SwitchFocus(),
SetScalingMode(), SetWindowConfig(). At the moment all calls are asynchronous because of a very
difficult to solve issue, that is when an application calls into the master to move a window, the
master calls into the application manager, but that in turn calls into the master again now. This
would block as the Fusion Dispatch thread is blocking on the call to the application manager and
cannot process the call from the application manager.

Add SaWManProcess flux including SetExiting() and RegisterManager().

Add permission management to throw functions.

Add SaWManWindow as an object that can be looked up, thrown, caught...

Change manager call to use FusionCall v3 copying back and forth the data instead of using
a shared memory pointer.