Core: Make Surface Buffer a FusionObject and implement locked surface resizing
The Surface Buffer object is created when a surface is created or reconfigured.

It has a global ref count of as long as it belongs to the surface.

dfb_surface_reconfig() no longer returns when a surface is locked, but decouples
the buffers by setting the back pointer from the Surface Buffer object to the surface to NULL
and decreasing the global ref count.

When a buffer is locked, the local ref count is increased, to keep the buffer
object alive until the buffer is unlocked.

Fixed locking for a few functions accessing surface members before locking it.

Weak 'locked' counter in buffer is now replaced by dfb_surface_buffer_locks()
function returning the local ref count of the Surface Buffer object.

All functions involved in unlocking and deallocating a buffer may no longer
use the surface back pointer unconditionally.