gfxcard / state: allow a graphics driver to hold extra state information
Two new callbacks are added to GraphicsDeviceFuncs, that if set, will be
called whenever a state is created or destroyed. This allows a graphics
driver to store custom state specific data in CardState, which is useful
if e.g. extra data needs to be preserved between calls to CheckState() and
SetState().