ImageProvider: Add secure IPC to (hardware) image providers
For secure fusion the hardware image providers need to be called
in the master process, as all other system and driver calls are.

This adds CoreDFB::CreateImageProvider() being passed a call id
from the slave for retrieving data buffer contents.

The new DataBuffer flux wraps that call, used by a master side
IDirectFBDataBuffer implementation. This is passed to the real
IDirectFBImageProvider instance.

The slave uses the new ImageProvider flux underneath a special
IDirectFBImageProvider implementation.