[valgrind] Use D_MAGIC_SET_ONLY when initializing the lock
In load_default_cursor a CoreSurfaceBufferLock is placed on the stack
and the addr is passed to dfb_surface_lock_buffer, this is calling
dfb_surface_buffer_lock_init and is using the D_MAGIC_SET on the lock.
The  macro will compare the current unitialized magic to the new one.

Use D_MAGIC_SET_ONLY to set the magic wihtout looking at the current
value, the alternative would be to initialize the CoreSurfaceBufferLock
on the stack with the magic set (or provide a macro for that).