VoodooPlay: Avoid broadcast discovery when client connects to an unknown player.
Add Voodoo messages for DISCOVER and SENDINFO as a temporary solution on the old
2323 port. Connections to different ports assume packet mode anyhow, now.

When connecting to 2323 without being able to lookup the player by its address
in the information collected by VoodooPlayer already, connect using RAW mode as
usual, but use the DISCOVER message and receive the SENDINFO from other player.
When this information reveals PACKET capability, reconnect with that mode.

With this commit the VoodooLink is no longer passed to the manager, but only
the VoodooConnection object being constructed beforehand now. The manager will
call connection->Start(this) at the end of its constructor.

Add WaitForData( timeout_ms ) to VoodooLink interface to prevent endless blocking
on the SENDINFO reply from old servers not supporting it.

Rename VPIF_LINK to VPIF_PACKET.

Change VOODOO_LINK_PORT (8676) to VOODOO_PLAY_PORT (8680). This is unused at
the moment, but should become the new standard Voodoo port.