One: Coma/One implementation
This is an alternative implementation of IComa and IComaComponent
to communicate via linux-one instead of linux-fusion.

You can select the new implementation by using "remote=%" for the
first process (creating the so called Name Service for the managers).

This will print the new QID of the Name Service, e.g.:

	(*) ComaOne/NameService: QID 1001

You can run additional processes by using "remote=%1001" in this case.

You can also run the name service with a predefined (static) ID, e.g.
using "remote=%=1" for the first and "remote=%1" for others. If the
Name Service is not started when another process in run, it will wait
until the Name Service is there.

There's no shared memory being used, only asynchronous One Queues!