SaWMan: Revert to old behaviour for SaWManProcess ref counting.
To fix a life cycle issue with SaWManWindow pointing to SaWManProcess
we increased the ref count of SaWManProcess per SaWManWindow. So the
SaWManProcess was only cleaned up when all the windows from that
process were destroyed.

The patch avoids this new behaviour and fixes the life cycle issue by
not having a pointer to SaWManProcess, but keeping a copy of relevant
values.