Core: Fix error path of surface allocation creation further, thanks to Rob McConnell!
"Please find attached a patch against the 1.6-pre code you have that fixes an assertion failing when the surface allocation object is successfully created, but the actual AllocateBuffer() function fails.  In this case, the surface allocation object destructor is called, but because the allocation size is not set the "CORE_SURFACE_ALLOCATION_ASSERT()" macro bails out.
 I've added a new state flag to indicate that the allocation is being initialised.  This ties the actual allocation and AllocateBuffer() functions together."