Core: Fix surface allocation decouple issue, thanks to Rob McConnell!
"This patch fixes the surface allocation decoupling when 'dfbtest_prealloc' is run as a slave and exits.  The wrong surface allocation object was being referenced in the call to 'dfb_surface_allocation_unlink()' in function 'dfb_surface_allocation_decouple()'.  In effect, it was trying to unlink the same surface allocation object twice and the assertion occurred in the call to 'dfb_surface_allocation_locks()'."