[PATCH] Fixed gfx driver module loading for slave processes.
Fixed dfb_gfxcard_load_driver in how graphics driver module  loading for slaves processes is handled to address the scenario  when there are multiple graphics drivers installed and the  module entry that should be loaded is not the first entry found.
It was found that a local reference to the next module link must  be stored prior to a call to the direct_module_unref function  because it will initialize the next field to NULL when there are  no more references to the module.  This is the same approach used  by the dfb_gfxcard_find_driver function used by the master  process.

Thanks to Timothy Strelchun <timothy.strelchun@intel.com>.