Core: Apply patch from Bryce Poole to fix a new deallocation bug in local memory surface pool in certain cases.
When the system align and pitch options are enabled, the system will allocate using posix_memalign otherwise
D_MALLOC is called. If in a debug build, D_MALLOC adds a description to the beginning of the memory and requires
D_FREE. The patch calls "free" for malloc'd surfaces and D_FREE for D_MALLOC's surfaces.