libdirect: Add String::CopyTLS(), D_String_CopyTLS() and D_String_PrintTLS().
CopyTLS allocates a slot in a thread local ring buffer of strings (char*),
copies the data and returns the pointer. This allows destruction of the
String object with subsequent code using the data only.

This is mainly meant for debugging purpose, so a ring buffer of up to 32 strings
was chosen that allows up to 32 such formatting prints as arguments to a debug
function like D_DEBUG_AT.