Task: Lots of updates to task manager code.
Add TaskLists for collecting a number of threads to run or running.
While the task is not flushed it can be added to such tasklists (increasing
the ref count on the task structure). The task structure will stay alive til
the last one released the task.

Keep allocations ref'ed in DisplayTask.

Add Flush() implementation in DisplayTask to add the task to the region's display tasks list,
removed from list in task Run() once the layer region got locked. This is the point til which
the layer shutdown code should be post poned.

Handle FPS as per task Finalise, not Run...

In dfb_layer_context_set_configuration() and other functions first wait for display task list to be empty.