drmkms: fix framebuffer allocation size
config.size.h has the real height of an image, so framebuffer size
calculation for packed color formats is done successfully.
However the framebuffer size in planar or semi-planar formats is
set incorrectly because the chroma plane size isn't added to the size.
This patch multiplies the pitch by drm_fake_height, which is takes into
account the size of the chroma plane.

Signed-off-by: Kazunori Kobayashi <kkobayas@igel.co.jp>