Task: Timed emits, optimisations, cleanups, few fixes
Implement time stamp based emit() of tasks when ts_emit is set and in
the future.

This is used by the DisplayTask implementation for layers that don't
support PTS with buffering of multiple display requests.

Set CSCAPS_DISPLAY_PTS if queueing of buffers with PTS is supported.

Fifo supports waiting with timeout based on target time stamp.