configure: update tslib version check
Since tslib 1.3 tslib's SONAME and pkg-config name and thus the ABI changed,
while the library is staying backwards compatible. The version number being
part of the name had been a mistake and is gone.

Since tslib-0 is well over 10 years old, we should not fall back to that
ancient version, but instead fall back to using a current version of tslib.

Signed-off-by: Martin Kepplinger <martin.kepplinger@ginzinger.com>